# app/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Weather API",
    "version": "1.0.0"
  },
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/weather": {
      "get": {
        "tags": ["Weather Routes"],
        "summary": "Retrieves weather",
        "description": "Retrieves the weather given a zipcode",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "zipCode",
            "in": "query",
            "required": true,
            "type": "string",
            "maxLength": 5,
            "description": "The zipcode to search the weather for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "temp": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/register": {
      "get": {
        "tags": ["User Routes"],
        "summary": "Create a user",
        "description": "Create a user in the backend",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "properties": {
                "userName": {
                  "type": "string",
                  "description": "The username of the user"
                },
                "firstName": {
                  "type": "string",
                  "description": "The first name of the user"
                },
                "lastName": {
                  "type": "string",
                  "description": "The last name of the user"
                },
                "age": {
                  "type": "integer",
                  "description": "The age of the user"
                }
              }
            },
            "description": "The user's details"
          }
        ],
        "responses": {
          "200": {
            "description": "200 response",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Error response",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}